refactor: 회원가입 API와 마이페이지 API 분리
